+++ bindings/matrix-sdk-ffi/src/client_builder.rs
@@ -78,17 +78,17 @@ pub enum HumanQrLoginError {
     Declined,
     #[error("An unknown error has happened.")]
     Unknown,
-    #[error("The QR code we scanned was not valid.")]
-    InvalidQrCode,
     #[error("The homeserver doesn't provide a sliding sync proxy in its configuration.")]
     SlidingSyncNotAvailable,
     #[error("Unable to use OIDC as the supplied client metadata is invalid.")]
     OidcMetadataInvalid,
+    #[error("The other device is not signed in and as such can't sign in other devices.")]
+    OtherDeviceNotSignedIn,
 }
 
 impl From<qrcode::QRCodeLoginError> for HumanQrLoginError {
     fn from(value: qrcode::QRCodeLoginError) -> Self {
-        use qrcode::QRCodeLoginError;
+        use qrcode::{QRCodeLoginError, SecureChannelError};
 
         match value {
             QRCodeLoginError::LoginFailure { reason, .. } => match reason {
@@ -97,6 +97,7 @@ impl From<qrcode::QRCodeLoginError> for HumanQrLoginError {
                 LoginFailureReason::UserCancelled => HumanQrLoginError::Cancelled,
                 _ => HumanQrLoginError::Unknown,
             },
+
             QRCodeLoginError::Oidc(e) => {
                 if let Some(e) = e.as_request_token_error() {
                     match e {
@@ -108,7 +109,18 @@ impl From<qrcode::QRCodeLoginError> for HumanQrLoginError {
                     HumanQrLoginError::Unknown
                 }
             }
-            QRCodeLoginError::SecureChannel(_) => HumanQrLoginError::ConnectionInsecure,
+
+            QRCodeLoginError::SecureChannel(e) => match e {
+                SecureChannelError::Utf8(_)
+                | SecureChannelError::MessageDecode(_)
+                | SecureChannelError::Json(_)
+                | SecureChannelError::RendezvousChannel(_) => HumanQrLoginError::Unknown,
+                SecureChannelError::SecureChannelMessage { .. }
+                | SecureChannelError::Ecies(_)
+                | SecureChannelError::InvalidCheckCode => HumanQrLoginError::ConnectionInsecure,
+                SecureChannelError::InvalidIntent => HumanQrLoginError::OtherDeviceNotSignedIn,
+            },
+
             QRCodeLoginError::UnexpectedMessage { .. }
             | QRCodeLoginError::CrossProcessRefreshLock(_)
             | QRCodeLoginError::DeviceKeyUpload(_)
@@ -480,7 +492,7 @@ impl ClientBuilder {
 
             Ok(client)
         } else {
-            Err(HumanQrLoginError::InvalidQrCode)
+            Err(HumanQrLoginError::OtherDeviceNotSignedIn)
         }
     }
 }