+++ bindings/matrix-sdk-ffi/src/event.rs
@@ -121,9 +121,9 @@ impl TryFrom<AnySyncStateEvent> for StateEventContent {
 pub enum MessageLikeEventContent {
     CallAnswer,
     CallInvite,
+    CallNotify { notify_type: NotifyType },
     CallHangup,
     CallCandidates,
-    CallNotify { notify_type: NotifyType },
     KeyVerificationReady,
     KeyVerificationStart,
     KeyVerificationCancel,
@@ -146,14 +146,14 @@ impl TryFrom<AnySyncMessageLikeEvent> for MessageLikeEventContent {
         let content = match value {
             AnySyncMessageLikeEvent::CallAnswer(_) => MessageLikeEventContent::CallAnswer,
             AnySyncMessageLikeEvent::CallInvite(_) => MessageLikeEventContent::CallInvite,
-            AnySyncMessageLikeEvent::CallHangup(_) => MessageLikeEventContent::CallHangup,
-            AnySyncMessageLikeEvent::CallCandidates(_) => MessageLikeEventContent::CallCandidates,
             AnySyncMessageLikeEvent::CallNotify(content) => {
                 let original_content = get_message_like_event_original_content(content)?;
                 MessageLikeEventContent::CallNotify {
                     notify_type: original_content.notify_type.into(),
                 }
             }
+            AnySyncMessageLikeEvent::CallHangup(_) => MessageLikeEventContent::CallHangup,
+            AnySyncMessageLikeEvent::CallCandidates(_) => MessageLikeEventContent::CallCandidates,
             AnySyncMessageLikeEvent::KeyVerificationReady(_) => {
                 MessageLikeEventContent::KeyVerificationReady
             }
@@ -289,6 +289,7 @@ pub enum MessageLikeEventType {
     CallCandidates,
     CallHangup,
     CallInvite,
+    CallNotify,
     KeyVerificationAccept,
     KeyVerificationCancel,
     KeyVerificationDone,
@@ -314,6 +315,7 @@ impl From<MessageLikeEventType> for ruma::events::MessageLikeEventType {
         match val {
             MessageLikeEventType::CallAnswer => Self::CallAnswer,
             MessageLikeEventType::CallInvite => Self::CallInvite,
+            MessageLikeEventType::CallNotify => Self::CallNotify,
             MessageLikeEventType::CallHangup => Self::CallHangup,
             MessageLikeEventType::CallCandidates => Self::CallCandidates,
             MessageLikeEventType::KeyVerificationReady => Self::KeyVerificationReady,