+++ crates/matrix-sdk/src/client/mod.rs
@@ -83,6 +83,7 @@ use crate::{
     matrix_auth::MatrixAuth,
     notification_settings::NotificationSettings,
     room_preview::RoomPreview,
+    send_queue::SendingQueueData,
     sync::{RoomUpdate, SyncResponse},
     Account, AuthApi, AuthSession, Error, Media, Pusher, RefreshTokenError, Result, Room,
     TransmissionProgress,
@@ -282,6 +283,11 @@ pub(crate) struct ClientInner {
     /// The verification state of our own device.
     #[cfg(feature = "e2e-encryption")]
     pub(crate) verification_state: SharedObservable<VerificationState>,
+
+    /// Data related to the [`SendingQueue`].
+    ///
+    /// [`SendingQueue`]: crate::send_queue::SendingQueue
+    pub(crate) sending_queue_data: Arc<SendingQueueData>,
 }
 
 impl ClientInner {
@@ -301,6 +307,7 @@ impl ClientInner {
         unstable_features: Option<BTreeMap<String, bool>>,
         respect_login_well_known: bool,
         event_cache: OnceCell<EventCache>,
+        sending_queue: Arc<SendingQueueData>,
         #[cfg(feature = "e2e-encryption")] encryption_settings: EncryptionSettings,
     ) -> Arc<Self> {
         let client = Self {
@@ -323,6 +330,7 @@ impl ClientInner {
             respect_login_well_known,
             sync_beat: event_listener::Event::new(),
             event_cache,
+            sending_queue_data: sending_queue,
             #[cfg(feature = "e2e-encryption")]
             e2ee: EncryptionData::new(encryption_settings),
             #[cfg(feature = "e2e-encryption")]
@@ -2102,6 +2110,7 @@ impl Client {
                 self.inner.unstable_features.get().cloned(),
                 self.inner.respect_login_well_known,
                 self.inner.event_cache.clone(),
+                self.inner.sending_queue_data.clone(),
                 #[cfg(feature = "e2e-encryption")]
                 self.inner.e2ee.encryption_settings,
             )
@@ -2149,7 +2158,6 @@ impl WeakClient {
     }
 
     /// Construct a [`WeakClient`] from a [`Client`].
-    #[cfg(test)]
     pub fn from_client(client: &Client) -> Self {
         Self::from_inner(&client.inner)
     }