+++ crates/matrix-sdk/src/lib.rs
@@ -51,6 +51,7 @@ pub mod pusher;
 pub mod room;
 pub mod room_directory_search;
 pub mod room_preview;
+pub mod send_queue;
 pub mod utils;
 pub mod futures {
     //! Named futures returned from methods on types in [the crate root][crate].